# api_geteway/api/docs/swagger.yaml
definitions:
...
  models.CreateStaffTarif:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      founded_at:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  models.CreateStaffTransaction:
    properties:
      amount:
        type: number
      information_about:
        type: string
      sale_id:
        type: string
      source_type:
        type: string
      staff_id:
        type: string
      transaction_type:
        type: string
    type: object
...
  models.StaffTarif:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      created_at:
        type: string
      founded_at:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.StaffTransaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      information_about:
        type: string
      sale_id:
        type: string
      source_type:
        type: string
      staff_id:
        type: string
      transaction_type:
        type: string
      updated_at:
        type: string
    type: object
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /tarifs:
    get:
      consumes:
      - application/json
      description: get Tarif
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        description: page of req
        in: query
        name: page
        required: true
        type: integer
      - description: filter by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StaffTarif'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: List Tarifs
      tags:
      - TARIF
    post:
      consumes:
      - application/json
      description: Create StsffTarif
      parameters:
      - description: tarif data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Create StaffTarif
      tags:
      - TARIF
  /tarifs/{id}:
    delete:
      consumes:
      - application/json
      description: delete tarif by Id
      parameters:
      - description: Tarif ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Delete By Id
      tags:
      - TARIF
    get:
      consumes:
      - application/json
      description: get staff tarif by ID
      parameters:
      - description: Tarif ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Get By Id
      tags:
      - TARIF
    put:
      consumes:
      - application/json
      description: api for update tarifs
      parameters:
      - description: id of staff tarif
        in: path
        name: id
        required: true
        type: string
      - description: data of tarif
        in: body
        name: tarif
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Update StaffTarif
      tags:
      - TARIF
  /transactions:
    get:
      consumes:
      - application/json
      description: get Transaction
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        description: page of req
        in: query
        name: page
        required: true
        type: integer
      - description: filter by type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StaffTransaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: List Transaction
      tags:
      - TRANSACTION
    post:
      consumes:
      - application/json
      description: Create Transaction
      parameters:
      - description: transaction data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Create Transaction
      tags:
      - TRANSACTION
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: delete transaction by Id
      parameters:
      - description: Transaction ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Delete By Id
      tags:
      - TRANSACTION
    get:
      consumes:
      - application/json
      description: get transaction by ID
      parameters:
      - description: Transaction ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Get By Id
      tags:
      - TRANSACTION
    put:
      consumes:
      - application/json
      description: api for update staffs transaction
      parameters:
      - description: id of transaction
        in: path
        name: id
        required: true
        type: string
      - description: data of transaction
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Update Transaction
      tags:
      - TRANSACTION
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"